refactor(config): remove old Settings loader, extract Tools to own file
Extracts the Tools struct from settings/mod.rs into settings/tools.rs
with deny_unknown_fields. Removes the Settings struct, load(), merge(),
load_global(), load_project(), load_file(), and all associated tests.

settings/mod.rs is now a thin re-export hub for the sub-type modules.
All production code uses CellaConfig exclusively.
